requirements.txt
elasticsearch==5.4.0
pika==0.11.0
pytest==3.2.3
pytest-cov==2.5.1
pyyaml==3.12
requests==2.18.4

